refactoring urls and packages hierarchy
