+++ ufcg/src/main/java/br/com/askufcg/controller/CommentController.java
@@ -2,7 +2,7 @@ package br.com.askufcg.controller;
 
 import br.com.askufcg.dto.PostCommentDTO;
 import br.com.askufcg.models.Comment;
-import br.com.askufcg.serviceImpl.CommentService;
+import br.com.askufcg.serviceImpl.comment.CommentService;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -15,12 +15,12 @@ public class CommentController {
         this.commentService = commentService;
     }
 
-    @PostMapping("/answer/{userId}/{answerId}")
+    @PostMapping("/answers/{answerId}/users/{userId}")
     public Comment addCommentAnswer(@RequestBody PostCommentDTO comment, @PathVariable("userId") Long userId, @PathVariable("answerId") Long answerid){
         return this.commentService.addCommentAnswer(comment, userId, answerid);
     }
 
-    @PostMapping("/question/{userId}/{questionId}")
+    @PostMapping("/questions/{questionId}/users/{userId}")
     public Comment addCommentQuestion(@RequestBody PostCommentDTO comment, @PathVariable("userId") Long userId, @PathVariable("questionId") Long questionId){
         return this.commentService.addCommentQuestion(comment, userId, questionId);
     }
